youtube-content-gen/resources/requirements.txt
youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
google-genai>=0.2.0
imageio-ffmpeg>=0.6.0
Pillow>=10.0.0
Jinja2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
